refac: componentizing simulator
